requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
bottleneck=1.3.5=py39h67323c0_0
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
brotlipy=0.7.0=py39h9ed2024_1003
bzip2=1.0.8=h1de35cc_0
ca-certificates=2022.10.11=hecd8cb5_0
certifi=2022.9.24=py39hecd8cb5_0
cffi=1.15.1=py39hc55c11b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=38.0.1=py39hf6deb26_0
cycler=0.11.0=pyhd3eb1b0_0
ffmpeg=4.2.2=h97e5cf8_0
fftw=3.3.9=h9ed2024_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=hd8bbffd_0
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
gmp=6.2.1=he9d5cce_3
gnutls=3.6.15=hed9c0bf_0
icu=58.2=h0a44026_3
idna=3.4=py39hecd8cb5_0
intel-openmp=2021.4.0=hecd8cb5_3538
joblib=1.1.1=py39hecd8cb5_0
jpeg=9e=hca72f7f_0
kiwisolver=1.4.2=py39he9d5cce_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libcxx=14.0.6=h9765a3e_0
libdeflate=1.8=h9ed2024_5
libffi=3.3=hb1e8313_2
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libiconv=1.16=hca72f7f_2
libidn2=2.3.2=h9ed2024_0
libopus=1.3.1=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libtasn1=4.16.0=h9ed2024_0
libtiff=4.4.0=h2cd0358_2
libunistring=0.9.10=h9ed2024_0
libvpx=1.7.0=h378b8a2_0
libwebp=1.2.4=h56c3ce4_0
libwebp-base=1.2.4=hca72f7f_0
libxml2=2.9.14=hbf8cd5e_0
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.3=h23ab428_1
matplotlib=3.5.3=py39hecd8cb5_0
matplotlib-base=3.5.3=py39hfb0c5b7_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
munkres=1.1.4=py_0
ncurses=6.3=hca72f7f_3
nettle=3.7.3=h230ac6f_1
numexpr=2.8.4=py39he696674_0
numpy=1.23.4=py39he696674_0
numpy-base=1.23.4=py39h9cd3388_0
openh264=2.1.1=h8346a28_0
openssl=1.1.1s=hca72f7f_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.5.1=py39h07fba90_0
pillow=9.2.0=py39hde71d04_1
pip=22.2.2=py39hecd8cb5_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py39hecd8cb5_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.15=hdfd78df_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.13.0=py3.9_0
pytz=2022.1=py39hecd8cb5_0
readline=8.2=hca72f7f_0
requests=2.28.1=py39hecd8cb5_0
scikit-learn=1.0.2=py39hae1ba45_1
scipy=1.9.3=py39h3d31255_0
setuptools=65.5.0=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.3=h707629a_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h5d9f67b_0
torch-summary=1.4.5=pypi_0
torchvision=0.14.0=py39_cpu
tornado=6.2=py39hca72f7f_0
typing_extensions=4.3.0=py39hecd8cb5_0
tzdata=2022f=h04d1e81_0
urllib3=1.26.12=py39hecd8cb5_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!157.20191217=h1de35cc_0
xz=5.2.6=hca72f7f_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.2=hcb37349_0
